test: Move dynamodb_local_environment fixture to universal/online_store/conftest.py
The fixture was defined in integration/online_store/conftest.py but its
only consumers were moved to universal/online_store/ — breaking fixture
scope. Move fixture and helper class to universal/online_store/conftest.py.
Strip integration/online_store/conftest.py (no remaining fixture users).

Also remove stale pytestmark = pytest.mark.integration from
test_hybrid_online_store.py (uses unittest.mock only, no real services).
